backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg==3.2.1
psycopg2-binary
python-dotenv==1.0.1
httpx==0.27.2